requirements.txt
altgraph==0.17.3
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
code128==0.3
colorama==0.4.5
cssselect2==0.7.0
deprecation==2.1.0
et-xmlfile==1.1.0
flake8==6.0.0
img2pdf==0.4.4
importlib-resources==5.4.0
jsonschema==4.17.3
lxml==4.9.2
macholib==1.16.2
mccabe==0.7.0
numpy==1.24.2
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pdf2image==1.16.3
pdf2img==0.1.2
pefile==2023.2.7
pikepdf==7.1.1
Pillow==9.4.0
poppler-utils==0.1.0
pycairo==1.23.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PyPDF3==1.0.6
pyproject-toml==0.0.10
PyQt6==6.4.2
PyQt6-Qt6==6.4.3
PyQt6-sip==13.4.1
pyrsistent==0.19.3
PySimpleGUI==4.60.4
python-barcode==0.14.0
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.0
shiboken6==6.4.3
six==1.16.0
tinycss2==1.2.1
toml==0.10.2
tqdm==4.64.1
urllib3==1.26.14
webencodings==0.5.1
xlrd==1.1.0
zipp==3.6.0
